organise the definition of the CLI app
