+++ cmd/swm/main.go
@@ -1,103 +1,60 @@
 package main
 
 import (
-	"errors"
 	"os"
 	"path"
 	"sort"
-	"strings"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	"go.i3wm.org/i3"
 	"gopkg.in/urfave/cli.v2"
 )
 
 var version = ""
+var app = &cli.App{
+	Name:    "swm",
+	Version: version,
+	Usage:   "swm <command>",
+	Flags: []cli.Flag{
+		&cli.StringFlag{Name: "profile", Usage: "The profile for the TMUX session", Value: getDefaultProfile()},
+		&cli.StringFlag{Name: "story", Usage: "The story for the TMUX session", Value: getDefaultStory()},
+		&cli.StringFlag{Name: "code-path", Usage: "The absolute path to the code path", Value: path.Join(os.Getenv("HOME"), "code")},
+		&cli.StringFlag{Name: "ignore-pattern", Usage: "The Regex pattern to ignore", Value: "^.snapshots$"},
+		&cli.BoolFlag{Name: "debug", Usage: "enable debug mode"},
+	},
+	Before: func(ctx *cli.Context) error {
+		// create the logger that pretty prints to the ctx.Writer
+		log.Logger = zerolog.New(zerolog.ConsoleWriter{Out: ctx.App.Writer}).
+			With().
+			Timestamp().
+			Str("ignore-pattern", ctx.String("ignore-pattern")).
+			Str("code-path", ctx.String("code-path")).
+			Str("profile", ctx.String("profile")).
+			Str("story", ctx.String("story")).
+			Logger().
+			Level(zerolog.InfoLevel)
+		// handle debug
+		if ctx.Bool("debug") {
+			log.Logger = log.Logger.Level(zerolog.DebugLevel)
+		}
 
-func main() {
-	app := &cli.App{
-		Name:    "swm",
-		Version: version,
-		Usage:   "swm <command>",
-		Flags: []cli.Flag{
-			&cli.StringFlag{Name: "profile", Usage: "The profile for the TMUX session", Value: getDefaultProfile()},
-			&cli.StringFlag{Name: "story", Usage: "The story for the TMUX session", Value: getDefaultStory()},
-			&cli.StringFlag{Name: "code-path", Usage: "The absolute path to the code path", Value: path.Join(os.Getenv("HOME"), "code")},
-			&cli.StringFlag{Name: "ignore-pattern", Usage: "The Regex pattern to ignore", Value: "^.snapshots$"},
-			&cli.BoolFlag{Name: "debug", Usage: "enable debug mode"},
+		return nil
+	},
+	Authors: []*cli.Author{
+		{
+			Name:  "Wael Nasreddine",
+			Email: "wael.nasreddine@gmail.com",
 		},
-		Before: func(ctx *cli.Context) error {
-			// create the logger that pretty prints to the ctx.Writer
-			log.Logger = zerolog.New(zerolog.ConsoleWriter{Out: ctx.App.Writer}).
-				With().
-				Timestamp().
-				Str("ignore-pattern", ctx.String("ignore-pattern")).
-				Str("code-path", ctx.String("code-path")).
-				Str("profile", ctx.String("profile")).
-				Str("story", ctx.String("story")).
-				Logger().
-				Level(zerolog.InfoLevel)
-			// handle debug
-			if ctx.Bool("debug") {
-				log.Logger = log.Logger.Level(zerolog.DebugLevel)
-			}
+	},
+	Commands: []*cli.Command{
+		// coder for code management
+		coderCmd,
+		// tmux for switch client
+		tmuxCmd,
+	},
+}
 
-			return nil
-		},
-		Authors: []*cli.Author{
-			{
-				Name:  "Wael Nasreddine",
-				Email: "wael.nasreddine@gmail.com",
-			},
-		},
-		Commands: []*cli.Command{
-			// coder for code management
-			{
-				Name: "coder",
-				Subcommands: []*cli.Command{
-					// add project
-					{
-						Name:      "add-project",
-						Usage:     "TODO",
-						Action:    coderAddProject,
-						ArgsUsage: "<url>",
-					},
-				},
-			},
-			// tmux for switch client
-			{
-				Name: "tmux",
-				Subcommands: []*cli.Command{
-					// switch client switches tmux client
-					{
-						Name:   "switch-client",
-						Usage:  "TODO",
-						Action: tmuxSwitchClient,
-						Flags: []cli.Flag{
-							&cli.BoolFlag{Name: "kill-pane", Usage: "kill the TMUX pane after switch client"},
-						},
-					},
-					// vim exit will save/exit any open vim
-					{
-						Name:   "vim-exit",
-						Usage:  "TODO",
-						Action: tmuxVimExit,
-					},
-					// kill-server will kill the server
-					{
-						Name:   "kill-server",
-						Usage:  "TODO",
-						Action: tmuxKillServer,
-						Flags: []cli.Flag{
-							&cli.BoolFlag{Name: "vim-exit", Usage: "if vim is found running, kill it"},
-						},
-					},
-				},
-			},
-		},
-	}
-
+func init() {
 	// sort the commands/flags
 	// TODO: refactor this to walk the tree rather than manually doing so
 	sort.Sort(cli.FlagsByName(app.Flags))
@@ -112,66 +69,11 @@ func main() {
 			}
 		}
 	}
+}
 
+func main() {
 	// run the app
 	if err := app.Run(os.Args); err != nil {
 		log.Fatal().Err(err).Msg("error occurred")
 	}
-}
-
-func getDefaultProfile() string {
-	var p string
-
-	tmuxSocketPath := os.Getenv("TMUX")
-	if tmuxSocketPath != "" {
-		profileStory := strings.Split(path.Base(tmuxSocketPath), ",")[0]
-		profileStoryArr := strings.Split(profileStory, "@")
-		if len(profileStoryArr) == 2 {
-			p = profileStoryArr[0]
-		}
-	}
-
-	if p == "" {
-		i3Workspace, err := getActiveI3WorkspaceName()
-		if err == nil && strings.Contains(i3Workspace, "@") {
-			p = strings.Split(i3Workspace, "@")[0]
-		}
-	}
-
-	return p
-}
-
-func getDefaultStory() string {
-	var s string
-
-	tmuxSocketPath := os.Getenv("TMUX")
-	if tmuxSocketPath != "" {
-		profileStory := strings.Split(path.Base(tmuxSocketPath), ",")[0]
-		profileStoryArr := strings.Split(profileStory, "@")
-		if len(profileStoryArr) == 2 {
-			s = profileStoryArr[1]
-		}
-	}
-	if s == "" {
-		i3Workspace, err := getActiveI3WorkspaceName()
-		if err == nil && strings.Contains(i3Workspace, "@") {
-			s = strings.Split(i3Workspace, "@")[1]
-		}
-	}
-
-	return s
-}
-
-func getActiveI3WorkspaceName() (string, error) {
-	// get the workspaces
-	workspaces, err := i3.GetWorkspaces()
-	if err != nil {
-		return "", err
-	}
-	for _, workspace := range workspaces {
-		if workspace.Focused {
-			return workspace.Name, nil
-		}
-	}
-	return "", errors.New("no active i3 workspace was found")
 }
